requirements.txt
langchain==0.0.340
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.18.1
openai>=1.1
faiss-cpu==1.7.4
altair==4
tiktoken==0.4.0
e2b
duckduckgo-search
wandb>=0.15.4
